chore: disable react deployment
We disabled the React bindings deployment for 24hs because after testing the
pipeline and making it work, we unpublished the "tests versions", but we
didn't know that you can't republish a package once unpublished for 24hs. And
since the "test versions" were the only versions, it unpublished the whole
package.